+++ lua/core/autocmds.lua
@@ -66,7 +66,7 @@ vim.api.nvim_create_autocmd("BufReadPost", {
    callback = function()
       local last_line = vim.fn.line([['"]])
       local total_lines = vim.fn.line("$")
-      if last_line > 1 and last_line <= total_lines then
+      if last_line >= 1 and last_line <= total_lines then
          vim.cmd('normal! g`"zz')
       end
    end,