+++ lua/plugins/dap.lua
@@ -42,7 +42,7 @@ return {
    {
       "jay-babu/mason-nvim-dap.nvim",
       event = "LspAttach",
-      dependencies = { "williamboman/mason.nvim", "mfussenegger/nvim-dap" },
+      dependencies = { "mason-org/mason.nvim", "mfussenegger/nvim-dap" },
       opts = {
          automatic_installation = true,
          ensure_installed = {