+++ lua/plugins/luarocks.lua
@@ -1,7 +1,0 @@
-return {
-   "vhyrro/luarocks.nvim",
-   -- Very high priority is required, luarocks.nvim should run as the first
-   -- plugin in your config.
-   priority = 2000,
-   config = true,
-}