+++ lua/plugins/treesitter.lua
@@ -2,7 +2,7 @@ return {
    -- Tree-sitter: LSP syntax tree and better syntax highlighting
    {
       "nvim-treesitter/nvim-treesitter",
-      -- TODO: migrate to 'main'
+      -- TODO: Migrate to 'main'
       tag = "v0.10.0", -- last version before master->main breaking changes
       event = { "BufReadPost", "BufNewFile" },
       cmd = { "TSInstall", "TSBufEnable", "TSBufDisable", "TSModuleInfo" },
@@ -14,6 +14,7 @@ return {
                "c",
                "lua",
                "luadoc",
+               "norg",
                "query",
                "vim",
                "vimdoc",